Refactor: Improve event handling and config management
This commit introduces two major refactorings to improve code quality and reduce duplication:

1.  Event Handling: All event handlers have been moved from `main.js` into a new `events` directory. Each event now has its own dedicated file, and a new `eventManager.js` is responsible for loading and subscribing them. This makes the event handling system more modular and easier to maintain.

2.  Configuration Management: A new generic `configManagerFactory.js` has been created to eliminate redundant code in the various config managers. All config managers (`configManager.js`, `kitsConfigManager.js`, `shopConfigManager.js`, and `ranksConfigManager.js`) now use this factory, which significantly reduces code duplication.
